data-streaming-service/requirements.txt
numpy
pandas
pytz
python-dotenv
tweepy==4.11.0
